refactor: Change postModal name to PostInputModal
